requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langchain-chroma
chromadb
pypdf
docx2txt
sqlalchemy
pyyaml
requests